Add files via upload
A Descriptive report on Forggith Pharmaceuticals, a pharmaceutical manufacturing company based in Germany. This is a two-page periodic report on both the sales and marketing performances spanning across four years to give an overview of the company's growth using the key performance indicators as the metrics in evaluating the anticipated growth target.