Subqueries using MySQL
An advanced way of data manipulation in SQL is using subqueries to filter datasets and output your results based on the type of filtering. Subqueries make use of two select queries in which the result of the outer select query in dependent on inner select query. In this project, some of the type of subqueries are used to show use cases of subqueries.