Database design and Entity-Relationship Diagram
In designing a database for accumulation of data in a relational database management system. The Database Administrator is tasked with specifying the kind of relationships that exist between the entities in the schema. It is quite paramount to visualize these relationships using an entity-relationship diagram (ERD). To further iterate the above explanation, I designed a database for an online movie rental (e.g Netflix) and I visualized the relationship between these entities in an Entity-Relationship Diagram. I used dbdiagram.io, an open source tool, to show the graphical representation.